requirements.txt
mcrcon
filelock
flask
kickpython
kickapi
google-api-python-client
google-auth-oauthlib
pyautogui
